requirements.txt
<<<<<<< HEAD
streamlit
deepface
tf-keras
numpy
pandas
schedule
python-dotenv
pillow
opencv-python-headless
twilio
=======
streamlit
deepface
tf-keras
numpy
pandas
schedule
python-dotenv
pillow
opencv-python-headless
twilio
>>>>>>> f56354bff23c57e9dbb309990488effecb6c4ad5
